Update systems to protocol implementations.
The framework for it is done. Systems are now protocol implementations,
making them easier to replace. Modules (and the systems as well) are
now less coupled to Containium. Configuration (the spec) is now a
system as well. Most systems still need to be updated to use this new
approach. Systems can now be Startable and Stoppable as well, which
is used within the simplified `with-systems` function.