Initial work on simplifying, and supporting swap, in module management
The `deploy`/`redeploy` command has been replaced with a single `activate`
command. This activate command now deploys, redeploys or swaps based on
the configuration and the current state of a module. The `undeploy`
command has been replaced with the `deactivate` command. An extra module
command is `describe`, which shows the known info about a module.