Logging over channels, capturing STDOUT. WIP.
This opens up the possibility of multiple (types of) console sessions
and command sessions. A log statement decides whether the messages
is of interest to only the command initiator, to only the console
sessions, or to both. Console sessions can tap onto the console mult,
whereas short-lived command sessions (which are automatically closed
after the command is done), are created by the initiator.